Add a new v1beta2 api for EventListener. This is just showing what the new API change might look like. It defines an annotation that we could use to distinguish between differing implementations.
Also, coalesces both KubernetesResource and CustomResources into
single Resources field in the Spec that's runtime.RawExtension.

Note that I'm not convinced that the field should be plural because
from what I can tell, it can only be a single resource. If we think
this should be (I really don't know) should be say an array, we could
revisit this. Left the name as is.